requirements.txt
Flask==3.0.0
requests==2.31.0
schedule==1.2.1
pywebview==4.4.1
